pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pyup"
authors = [{name = "Alyetama", email = "56323389+Alyetama@users.noreply.github.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "paramiko~=2.10.2",
    "pymongo[srv]~=4.0.2",
    "pyperclip~=1.8.2",
    "PyYAML~=6.0",
    "loguru~=0.5.3",
    "tqdm~=4.62.3",
    "rich~=10.11.0"
]

[project.urls]
Home = "https://github.comm/Alyetama/pyup"

[project.scripts]
pyup = "pyup:run.main"
